Implement simple, monomorphic type checking.
Currently only Int64, Bool, and Unit types are supported. There are
many things in here that will need to change (a lot) when a Real Type
System comes to town, but for the moment this works and seems to be
correct. In particular, code generation will need to be refactored to
be more type-flexible, when we have more than just integral values.